pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cicada"
version = "0.1.0"
authors = [
    { name = "Oaklight", email = "oaklight@gmx.com" }, # Updated to match setup.py
]
description = "A Python package for the Cicada project"
readme = "README_en.md" # Updated to match setup.py
requires-python = ">=3.11"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests==2.32.3",
    "pyyaml==6.0.2",
    "tenacity==9.0.0",
    "cascadio==0.0.14",
    "gmsh==4.13.1",
    "plyfile==1.1",
    "trimesh==4.5.1",
    "pyglet<2",
    "pyrender==0.1.45",
    "openai==1.55.3",
    "build123d==0.8.0",
    "flake8==7.1.1",
    "psycopg2-binary==2.9.10",
    "sqlalchemy==2.0.36",
    "pgvector==0.3.6",
    "sqlite-vec==0.1.6",
    "questionary==2.1.0",
    "blessed==1.20.0",
]

[project.urls]
Homepage = "https://github.com/Oaklight/cicada" # Updated to match setup.py
